// src/types/content_unit.ts
export type Content_unit = {
    id: String
    course_id : String
    unit_no : Number
    title: String
    keyword: String
    language: String
    content_type: String
    content_name: String
    content_url: String
    uploaded_by: String
    upload_time: Date
    status: Boolean
    deleted: Boolean
    created: Date
  }